Removed intermediate class
