+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/Services/EndpointExecuters/AbstractEndpointExecuter.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace GraphQLAPI\GraphQLAPI\Services\EndpointExecuters;
 
 use GraphQLAPI\GraphQLAPI\Registries\ModuleRegistryInterface;
+use GraphQLAPI\GraphQLAPI\Services\CustomPostTypes\AbstractGraphQLEndpointCustomPostType;
 use PoP\ComponentModel\Instances\InstanceManagerInterface;
 
 abstract class AbstractEndpointExecuter implements EndpointExecuterInterface
@@ -29,6 +30,21 @@ abstract class AbstractEndpointExecuter implements EndpointExecuterInterface
         if ($enablingModule !== null) {
             return $this->moduleRegistry->isModuleEnabled($enablingModule);
         }
+
+        // Check we're loading the corresponding CPT
+        $customPostType = $this->getCustomPostType();
+        if (!\is_singular($customPostType->getCustomPostType())) {
+            return false;
+        }
+
+        // Check the endpoint is not disabled
+        global $post;
+        if (!$customPostType->isEndpointEnabled($post)) {
+            return false;
+        }
+        
         return true;
     }
+    
+    abstract protected function getCustomPostType(): AbstractGraphQLEndpointCustomPostType;
 }
